+++ src/SlackConnector.Tests.Unit/Connections/Clients/HandshakeClientTests.cs
@@ -4,6 +4,7 @@ using RestSharp;
 using Should;
 using SlackConnector.Connections;
 using SlackConnector.Connections.Clients;
+using SlackConnector.Connections.Clients.Handshake;
 using SlackConnector.Connections.Responses;
 using SlackConnector.Tests.Unit.Stubs;
 using SpecsFor;
