+++ src/SlackConnector/Connections/Clients/Channel/IChannelClient.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace SlackConnector.Connections.Clients.Channel
+{
+    internal interface IChannelClient
+    {
+        Task<Models.Channel> JoinDirectMessageChannel(string slackKey, string user);
+    }
+}
