+++ src/SlackConnector/SlackConnector.csproj
@@ -58,12 +58,12 @@
     <Compile Include="BotHelpers\IMentionDetector.cs" />
     <Compile Include="BotHelpers\IChatHubInterpreter.cs" />
     <Compile Include="Connections\ConnectionFactory.cs" />
-    <Compile Include="Connections\Clients\HandshakeClient.cs" />
-    <Compile Include="Connections\Clients\IHandshakeClient.cs" />
-    <Compile Include="Connections\Clients\ChannelClient.cs" />
-    <Compile Include="Connections\Clients\IChannelClient.cs" />
-    <Compile Include="Connections\Clients\IChatClient.cs" />
-    <Compile Include="Connections\Clients\ChatClient.cs" />
+    <Compile Include="Connections\Clients\Handshake\HandshakeClient.cs" />
+    <Compile Include="Connections\Clients\Handshake\IHandshakeClient.cs" />
+    <Compile Include="Connections\Clients\Channel\ChannelClient.cs" />
+    <Compile Include="Connections\Clients\Channel\IChannelClient.cs" />
+    <Compile Include="Connections\Clients\Chat\IChatClient.cs" />
+    <Compile Include="Connections\Clients\Chat\ChatClient.cs" />
     <Compile Include="Connections\Clients\IResponseVerifier.cs" />
     <Compile Include="Connections\Clients\ResponseVerifier.cs" />
     <Compile Include="Connections\Models\Group.cs" />
